+++ modules/Students/includes/Address.inc.php
@@ -9,18 +9,15 @@ if ( isset( $_POST['day_values'] )
 	&& isset( $_POST['month_values'] )
 	&& isset( $_POST['year_values'] ) )
 {
-	foreach( (array)$_REQUEST['month_values'] as $field_category => $columns )
-	{
-		foreach( $columns as $column => $month )
-		{
-			$_REQUEST['values'][$field_category][$column] =
-			$_POST['values'][$field_category][$column] = RequestedDate(
-				$_REQUEST['day_values'][$field_category][$column],
-				$month,
-				$_REQUEST['year_values'][$field_category][$column]
-			);
-		}
-	}
+	$requested_dates = RequestedDates(
+		$_REQUEST['day_values'],
+		$_REQUEST['month_values'],
+		$_REQUEST['year_values']
+	);
+
+	$_REQUEST['values'] = array_merge_recursive( $_REQUEST['values'], $requested_dates );
+
+	$_POST['values'] = array_merge_recursive( $_POST['values'], $requested_dates );
 }
 
 if ( isset( $_POST['values'] )